+++ src/lib/query/proposal.ts
@@ -1,34 +1,5 @@
 import { graphql } from "lib/gql";
 
-export const getRelatedProposalsByContractAddressPagination = graphql(`
-  query getRelatedProposalsByContractAddressPagination(
-    $contractAddress: String!
-    $offset: Int!
-    $pageSize: Int!
-  ) {
-    contract_proposals(
-      where: { contract: { address: { _eq: $contractAddress } } }
-      order_by: { proposal_id: desc }
-      offset: $offset
-      limit: $pageSize
-    ) {
-      proposal {
-        title
-        status
-        voting_end_time
-        deposit_end_time
-        type
-        account {
-          address
-        }
-        is_expedited
-      }
-      proposal_id
-      resolved_height
-    }
-  }
-`);
-
 export const getRelatedProposalsCountByContractAddress = graphql(`
   query getRelatedProposalsCountByContractAddress($contractAddress: String!) {
     contract_proposals_aggregate(